blogapi/requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
